UtilitiesCalculator/docker/requirements.txt
bcrypt==3.2.0
blinker==1.4
cffi==1.14.3
click==7.1.2
DateTime==4.3
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==1.3.13
pdfkit==0.6.1
psycopg2-binary==2.8.4
pycparser==2.20
pytz==2020.1
python-dotenv==0.14.0
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
WTForms==2.3.3
WTForms-SQLAlchemy==0.2
zope.interface==5.1.2
